+++ pyanalyze/test_signature.py
@@ -1125,7 +1125,7 @@ class TestOverload(TestNameCheckVisitorBase):
             assert_is_value(f.__round__(), TypedValue(int))
             assert_is_value(f.__round__(None), TypedValue(int))
             f.__round__(ndigits=None)  # E: incompatible_call
-            assert_is_value(f.__round__(1), TypedValue(float))
+            assert_is_value(f.__round__(1), TypedValue(float) | TypedValue(int))
 
     @assert_passes()
     def test_runtime(self):