+++ pyanalyze/type_object.py
@@ -66,13 +66,6 @@ class TypeObject:
             self.is_universally_assignable = issubclass(self.typ, mock.NonCallableMock)
         self.is_thrift_enum = hasattr(self.typ, "_VALUES_TO_NAMES")
         self.base_classes |= set(get_mro(self.typ))
-        # As a special case, the Python type system treats int as
-        # a subtype of float, and both int and float as subtypes of complex.
-        if self.typ is int or safe_in(int, self.base_classes):
-            self.artificial_bases.add(float)
-            self.artificial_bases.add(complex)
-        if self.typ is float or safe_in(float, self.base_classes):
-            self.artificial_bases.add(complex)
         if self.is_thrift_enum:
             self.artificial_bases.add(int)
         self.base_classes |= self.artificial_bases